requirements.txt
aiogram==3.15.0
sqlalchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
loguru
alembic
